Fixed torch burning ticks progressing while the game is paused.
Every frame, if the game is not paused, the torch checks if more than 20 seconds has elapsed since the last tick, and if so, reduces condition. This check is based on real time­. Therefore, if a tick occurs and the player pauses immediately for at least 20 seconds, once they come back to gameplay, the torch will burn again immediately. No game mechanic should progress while the game is paused.

This implementation instead uses a time buffer that increases in time while the game is not paused (and the torch is active of course). One potential issue is that the timer does not reset when unequipping the torch, but it might be a good thing, since it prevents "exploits" based on unequipping the torch every 19 seconds.